% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NT.R
\name{mean_nt}
\alias{mean_nt}
\title{Calculo de noches tropicales para un SpatRaster}
\usage{
mean_nt(raster_entrada, year_inicio, year_fin)
}
\arguments{
\item{raster_entrada}{un objeto de la clase SpatRaster con las temperaturas mínimas por dia}

\item{year_inicio}{año de inicio del periodo que se desea calcular}

\item{year_fin}{año de inicio del periodo que se desea calcular}
}
\value{
devuelve un objeto de la clase SpatRaster con la media anual de noches tropicales por pixel
}
\description{
\code{mean_nt} devuelve las noches tropicales promedio anuales
}
\examples{
\dontrun{
library(terra)
r_base <- rast(nrows = 10, ncols = 10,
            xmin = 0, xmax = 10,
            ymin = 0, ymax = 10)

r1 <- rast(r_base)
r2 <- rast(r_base)

values(r1) <- runif(ncell(r1), min = 10, max = 30)
values(r2) <- runif(ncell(r2), min = 10, max = 30)

tm_min <- c(r1, r2)

time(tm_min) <- as.Date(c("2001-01-01", "2002-01-01"))

tm_min_mean <- mean_nt(tm_min,2001,2002)}

}
